apps/websocket/requirements.txt
websockets==12.0
redis==5.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
structlog==23.2.0 